requirements.txt
django
opencv-python
jupyter
scikit-learn
scikit-image
Pillow
numpy 
matplotlib
pandas
